Working now with Element record instead of struct.
This is fine in isolation, but not as good for interoperability. I would be much happier if Element was defined in clojure.xml. I'm also not a fan of defining types for data unless you have to, but in this case I feed like I had to - consider the substitution ${x} in a template. I need to know whether x should be stringified, or treated as a chunk of XML. Previously I was doing that with (and (map? x) (:tag x)). Checking for Element is much cleaner, but means than non-Element XML (eg: returned from clojure.xml/parse) will not be treated as XML. Need to think about this some more, but for now it works as-is.